move `create_summary` method to results_base
